backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy
psycopg2-binary
stripe
slowapi
reportlab
httpx
PyMuPDF
anthropic
cryptography
python-dotenv
